+++ R/st_nn.R
@@ -15,6 +15,7 @@
 #' @param maxdist Search radius (\strong{in meters}). Points farther than search radius are not considered. Default is \code{Inf} meaning that search is unconstrained
 #' @param returnDist \code{logical}; whether to return a second \code{list} with the distances between detected neighbors.
 #' @param progress Display progress bar? (default \code{TRUE}
+#' @param parallel Number of parallel processes or a predefined socket cluster. With \code{parallel=1} uses ordinary, non-parallel processing. Parallel processing is done with the \code{parallel} package
 #' @return  \itemize{
 #' \item{If \code{sparse=TRUE} (the default), a sparse \code{list} with list element \code{i} being a numeric vector with the indices \code{j} of neighboring features from \code{y} for the feature \code{x[i,]}, or \code{integer(0)} in case there are no neighbors.}
 #' \item{If \code{sparse=FALSE}, a \code{logical} matrix with element \code{[i,j]} being \code{TRUE} when \code{y[j,]} is a neighbor of \code{x[i]}.}
@@ -90,7 +91,7 @@
 #'
 #' # Large example - Polygons
 #' set.seed(1)
-#' n = 50
+#' n = 150
 #' x = data.frame(
 #'   lon = (runif(n) * 2 - 1) * 70,
 #'   lat = (runif(n) * 2 - 1) * 70
@@ -99,14 +100,22 @@
 #' x = st_transform(x, 32630)
 #' x = st_buffer(x, 1000000)
 #' start = Sys.time()
-#' result = st_nn(x, x, k = 3)
-#' result[1:3]
+#' result1 = st_nn(x, x, k = 3)
+#' result1[1:3]
 #' end = Sys.time()
 #' end - start
 #'
+#' # Large example - Polygons - Parallel processing
+#' start = Sys.time()
+#' result2 = st_nn(x, x, k = 3, parallel = 4)
+#' result2[1:3]
+#' end = Sys.time()
+#' end - start
+#' all.equal(result1, result2)
+#'
 #' }
 
-st_nn = function(x, y, sparse = TRUE, k = 1, maxdist = Inf, returnDist = FALSE, progress = TRUE) {
+st_nn = function(x, y, sparse = TRUE, k = 1, maxdist = Inf, returnDist = FALSE, progress = TRUE, parallel = 1) {
 
   # To geometry
   x = sf::st_geometry(x)
@@ -124,23 +133,82 @@ st_nn = function(x, y, sparse = TRUE, k = 1, maxdist = Inf, returnDist = FALSE, 
   if(!is.na(sf::st_crs(x)) & !is.na(sf::st_crs(y)) & sf::st_crs(x) != sf::st_crs(y))
     stop("'x' and 'y' need to be in the same CRS")
 
-  # Determine geometry type and projection
-  if(!class(x)[1] == "sfc_POINT" | !class(y)[1] == "sfc_POINT") {
-    message("lines or polygons")
-    result = .st_nn_poly(x, y, k, maxdist, progress)
-  } else {
-    if(!is.na(sf::st_crs(x)) & !is.na(sf::st_crs(y)) & sf::st_is_longlat(x) & sf::st_is_longlat(y)) {
-      message("lon-lat points")
-      result = .st_nn_pnt_geo(x, y, k, maxdist, progress)
+  # Determine geometry type and projection & calculate IDs+dists
+  # Single thread
+  if(parallel == 1) {
+    if(!class(x)[1] == "sfc_POINT" | !class(y)[1] == "sfc_POINT") {
+      message("lines or polygons")
+      result = .st_nn_poly(x, y, k, maxdist, progress)
     } else {
-      message("projected points")
-      result = .st_nn_pnt_proj(x, y, k, maxdist, progress)
+      if(!is.na(sf::st_crs(x)) & !is.na(sf::st_crs(y)) & sf::st_is_longlat(x) & sf::st_is_longlat(y)) {
+        message("lon-lat points")
+        result = .st_nn_pnt_geo(x, y, k, maxdist, progress)
+      } else {
+        message("projected points")
+        result = .st_nn_pnt_proj(x, y, k, maxdist, progress)
+      }
     }
+    # Returned sparse lists
+    ids = result[[1]]
+    dists = result[[2]]
   }
-
-  # Returned sparse lists
-  ids = result[[1]]
-  dists = result[[2]]
+  # Parallel
+  if(parallel > 1) {
+    if(!class(x)[1] == "sfc_POINT" | !class(y)[1] == "sfc_POINT") {
+      message("lines or polygons")
+      if(.Platform$OS.type == "unix") {
+        result = parallel::mclapply(
+          split(x, 1:length(x)),
+          function(i) .st_nn_poly(i, y, k, maxdist, progress = FALSE),
+          mc.cores = parallel
+        )
+      } else {
+        result = parallel::parLapply(
+          parallel::makeCluster(parallel),
+          split(x, 1:length(x)),
+          function(i) .st_nn_poly(i, y, k, maxdist, progress = FALSE)
+        )
+      }
+    } else {
+      if(!is.na(sf::st_crs(x)) & !is.na(sf::st_crs(y)) & sf::st_is_longlat(x) & sf::st_is_longlat(y)) {
+        message("lon-lat points")
+        if(.Platform$OS.type == "unix") {
+          result = parallel::mclapply(
+            split(x, 1:length(x)),
+            function(i) .st_nn_pnt_geo(i, y, k, maxdist, progress = FALSE),
+            mc.cores = parallel
+          )
+        } else {
+          result = parallel::parLapply(
+            parallel::makeCluster(parallel),
+            split(x, 1:length(x)),
+            function(i) .st_nn_pnt_geo(i, y, k, maxdist, progress = FALSE)
+          )
+        }
+      } else {
+        message("projected points")
+        if(.Platform$OS.type == "unix") {
+          result = parallel::mclapply(
+            split(x, 1:length(x)),
+            function(i) .st_nn_pnt_proj(i, y, k, maxdist, progress = FALSE),
+            mc.cores = parallel
+          )
+        } else {
+          result = parallel::parLapply(
+            parallel::makeCluster(parallel),
+            split(x, 1:length(x)),
+            function(i) .st_nn_pnt_proj(i, y, k, maxdist, progress = FALSE)
+          )
+        }
+      }
+    }
+    ids = lapply(result, function(i) i[[1]])
+    ids = do.call(c, ids)
+    names(ids) = NULL
+    dists = lapply(result, function(i) i[[1]])
+    dists = do.call(c, dists)
+    names(dists) = NULL
+  }
 
   # To dense matrix?
   if(!sparse) {